+++ app/assets/stylesheets/application.scss
@@ -10,12 +10,3 @@
 // Your CSS partials
 @import "components/index";
 @import "pages/index";
-
-body {
-	max-width: 1024px;
-	margin: 0 auto; // Optional: Center the content horizontally
-  }
-
-  .popup-link {
-	cursor: pointer;
-  }
