--- app/javascript/controllers/search_campaigns_controller.js
@@ -1,25 +1,0 @@
-import { Controller } from "@hotwired/stimulus"
-
-export default class extends Controller {
-  static targets = ["form", "input", "list"]
-
-  connect() {
-    console.log(this.formTarget)
-    console.log(this.inputTarget)
-    console.log(this.listTarget)
-  }
-
-  update() {
-    console.log(this.formTarget);
-    console.log(this.formTarget.action);
-    
-    const url = `${this.formTarget.action}?query=${this.inputTarget.value}`
-    fetch(url, {headers: {"Accept": "text/plain"}})
-    // i am expecting a response type text, not html
-    // not expecting a html response, want just the text format
-      .then(response => response.text())
-      .then((data) => {
-        this.listTarget.outerHTML = data
-      })
-  }
-}
