+++ app/views/campaigns/index.html.erb
@@ -56,7 +56,7 @@
   </div>  
   <div class="mt-4 position-absolute">
     <div class="">
-      <button class="btn btn-secondary" style="height: 80px; width: 250px" type="submit">Create New</button> <button class="btn btn-secondary" style="height: 80px; width: 250px" type="submit">Load More</button>
+      <button class="btn btn-secondary" style="height: 80px; width: 250px" type="submit">Create New</button>
     </div>
   </div>
 </div>
