+++ app/views/devise/sessions/new.html.erb
@@ -1,7 +1,7 @@
-<div style="margin-top: 280px">
+<div class="background" style="height: 94vh;">
   <div id="screen-size" class="d-flex justify-content-center">
-    <div class="row m-5 p-5 box" style="width: 500px"> 
-      <h1 class="mb-5 text-center text-deark">Welcome Back</h1>
+    <div class="row m-5 p-5 box bg-white shadow-lg welcome-height" style="width: 500px"> 
+      <h1 class="mb-5 text-center text-dark">Welcome Back</h1>
           <h3  class="mb-3 text-left">Log in</h3>
             <%= simple_form_for(resource, as: resource_name, url: session_path(resource_name)) do |f| %>
               <div class="form-inputs text-primary">
