--- app/views/leads/show.html.erb
@@ -1,1 +1,0 @@
-showw!!
