+++ app/views/pages/dashboard.html.erb
@@ -3,7 +3,7 @@
 	<div class="container">
 		<div class="row">
 			<div class="col">
-				<div class="container p-3 border border-dark">
+				<div class="container p-3 box shadow-sm">
 					<h3> TOTAL LEADS </h3>
 					<p> THIS WEEK </p>
 					<h2><strong>40</strong></h2>
@@ -11,7 +11,7 @@
 			</div>
 
 			<div class="col">
-				<div class="container p-3 border border-dark">
+				<div class="container p-3 box shadow-sm">
 					<h3> QUALIFIED LEADS </h3>
 					<p> THIS WEEK </p>
 					<h2><strong>40</strong></h2>
@@ -19,7 +19,7 @@
 			</div>
 
 			<div class="col">
-				<div class="container p-3 border border-dark">
+				<div class="container p-3 box shadow-sm">
 					<h3> SPAM LEADS </h3>
 					<p> THIS WEEK </p>
 					<h2><strong>40</strong></h2>
@@ -29,7 +29,7 @@
 	</div>
 
 	<div class="mt-5">
-		<div class="container p-3 border border-dark" style="width: 915px;">
+		<div class="container p-3 box shadow-sm" style="width: 915px;">
 			<h2> GRAPH </h2>
 			<h4>MONTHLY LEADS</h4>
