+++ config/environments/development.rb
@@ -58,7 +58,7 @@ Rails.application.configure do
 
   # Suppress logger output for asset requests.
   config.assets.quiet = true
-
+  config.assets.debug = true
   # Raises error for missing translations.
   # config.i18n.raise_on_missing_translations = true
